backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
django-redis==5.4.0
tabulate==0.9.0
pillow==10.2.0
boto3==1.34.17
celery==5.4.0
langchain==0.1.16
langchain-experimental==0.0.57
langchain-openai==0.1.3
langchain-anthropic==0.1.11
pypdf==4.2.0
openai==1.23.6
qdrant-client==1.9.0
drf-spectacular==0.27.2
django-phonenumber-field[phonenumberslite]==7.3.0
channels==4.0.0
channels-redis==4.2.0
daphne==4.1.0
beautifulsoup4==4.12.3
html2text==2024.2.26
dj-rest-auth==2.2.7
django-allauth==0.52.0
md2pdf==1.0.1
pdf2docx==0.5.8
django-storages[google]